validations

+++ src/app/pokemons/pokemons-list/pokemons-list.component.ts
@@ -10,12 +10,13 @@ declare const $: any;
   styleUrls: ['./pokemons-list.component.css']
 })
 export class PokemonsListComponent implements OnInit {
-  pokemons:IPokemon[] = []
+  pokemons: IPokemon[] = []
+  isInvalid:boolean = false
 
   favoriteForm: FormGroup = new FormGroup({
-    name: new FormControl('',Validators.required),
-    alias: new FormControl(''),
-    createdAt: new FormControl(new Date(),Validators.required),
+    name: new FormControl('', Validators.required),
+    alias: new FormControl('', Validators.required),
+    createdAt: new FormControl(new Date(), Validators.required),
   });
 
   constructor(private _pokemonService: PokemonsServiceService) { }
@@ -25,68 +26,80 @@ export class PokemonsListComponent implements OnInit {
   }
 
   getData() {
-    this._pokemonService.getPokemonList().subscribe((data:any) =>{
+    this._pokemonService.getPokemonList().subscribe((data: any) => {
       this.pokemons = data.results;
 
       this.getFavorites();
-      
+
     })
   }
 
 
-  getFavorites(){
+  getFavorites() {
     let values = this._pokemonService.getFavoritesList();
-      
-      this.pokemons.forEach((element)=>{
-        element.isFavorite = false;
-        values.forEach((session)=>{
-          if(element.name == session.name){
-            element.isFavorite = true;
-          }
-        })
+
+    this.pokemons.forEach((element) => {
+      element.isFavorite = false;
+      values.forEach((session) => {
+        if (element.name == session.name) {
+          element.isFavorite = true;
+        }
       })
+    })
   }
 
+ 
 
-  deleteFavorite(item){
+  deleteFavorite(item) {
 
     this.favoriteForm.patchValue(item);
 
-    let response = this._pokemonService.deleteFromFavorites(this.favoriteForm.controls.name.value);
+    if (confirm('Vas a liberar este pokemon') == true) {
+      let response = this._pokemonService.deleteFromFavorites(this.favoriteForm.controls.name.value);
 
-    if(response){
-      this.getFavorites();
-      alert('Pokemon eliminado')
-      
+      if (response) {
+        this.getFavorites();
+        alert('Pokemon eliminado')
 
-    }else{
-      console.log('Pokemon no se pudo eliminar')
+
+      } else {
+        console.log('Pokemon no se pudo eliminar')
+      }
     }
+
   }
 
-  setFavorite(item){
-   this.favoriteForm.patchValue(item);
-   console.log(item)
+  setFavorite(item) {
+    this.favoriteForm.patchValue(item);
+    console.log(item)
   }
-  
-  
-  saveFavorite(){
+
+
+  saveFavorite() {
     console.log(this.favoriteForm.value)
-    
-    if(!this.favoriteForm.valid){
+
+    if (!this.favoriteForm.valid) {
+      this.isInvalid = true
       return;
     }
- 
+
+    this.isInvalid = false
     let response = this._pokemonService.addToFavorites(this.favoriteForm.value);
- 
-     if(response){
-       $('#addFavoritePokemon').modal('hide');
-       this.getFavorites();
-       alert('Pokemon Agregado')
- 
-     }else{
-       console.log('Pokemon no se pudo agregar')
-     }
+
+    if (response) {
+      this.resetForm()
+      this.getFavorites();
+      alert('Pokemon Agregado')
+
+    } else {
+      console.log('Pokemon no se pudo agregar')
+    }
   }
+
+  resetForm(){
+    this.isInvalid = false
+    this.favoriteForm.reset()
+  }
+
 
 }
